refactor single module
