+++ lib/zakuro/calculation/summary/single.rb
@@ -6,6 +6,8 @@ require_relative '../range/operated_range'
 
 require_relative '../range/full_range'
 
+require_relative './internal/operation'
+
 # :nodoc:
 module Zakuro
   # :nodoc:
@@ -25,107 +27,83 @@ module Zakuro
         # @return [Result::Single] 一日検索結果（和暦日）
         #
         def self.get(context:, date: Western::Calendar.new)
-          full_range = Calculation::Range::FullRange.new(context: context, start_date: date)
-          years = full_range.get
+          years = get_full_range_years(context: context, date: date)
 
+          data = get_data(context: context, years: years, date: date)
+
+          operation = get_operation(years: years, date: date)
+
+          Result::Single.new(
+            data: data,
+            operation: operation
+          )
+        end
+
+        #
+        # 完全範囲を取得する
+        #
+        # @param [Context] context 暦コンテキスト
+        # @param [Western::Calendar] date 西暦日
+        #
+        # @return [Array<Base::Year>] 完全範囲
+        #
+        def self.get_full_range_years(context:, date: Western::Calendar.new)
+          full_range = Calculation::Range::FullRange.new(context: context, start_date: date)
+          full_range.get
+        end
+        private_class_method :get_full_range_years
+
+        #
+        # 運用結果範囲を取得する
+        #
+        # @param [Context] context 暦コンテキスト
+        # @param [Array<Base::Year>] years 完全範囲
+        # @param [Western::Calendar] date 西暦日
+        #
+        # @return [Array<Base::OperatedYear>] 運用結果範囲
+        #
+        def self.get_operated_range_years(context:, years:, date: Western::Calendar.new)
+          operated_range = Calculation::Range::OperatedRange.new(
+            context: context, start_date: date, years: years
+          )
+          operated_range.get
+        end
+        private_class_method :get_operated_range_years
+
+        #
+        # 1日を取得する
+        #
+        # @param [Context] context 暦コンテキスト
+        # @param [Array<Base::Year>] years 完全範囲
+        # @param [Western::Calendar] date 西暦日
+        #
+        # @return [Data::SingleDay] 1日
+        #
+        def self.get_data(context:, years:, date: Western::Calendar.new)
+          operated_years = get_operated_range_years(context: context, years: years, date: date)
+
+          Calculation::Specifier::SingleDay.get(
+            years: operated_years, date: date
+          )
+        end
+        private_class_method :get_data
+
+        #
+        # 完全範囲を取得する
+        #
+        # @param [Context] context 暦コンテキスト
+        # @param [Western::Calendar] date 西暦日
+        #
+        # @return [Array<Base::Year>] 完全範囲
+        #
+        def self.get_operation(years:, date: Western::Calendar.new)
           calc_date = Calculation::Specifier::SingleDay.get(
             years: years, date: date
           )
 
-          operated_range = Calculation::Range::OperatedRange.new(
-            context: context, start_date: date, years: years
-          )
-          operated_years = operated_range.get
-
-          Result::Single.new(
-            data: Calculation::Specifier::SingleDay.get(
-              years: operated_years, date: date
-            ),
-            operation: create_operation(calc_date: calc_date)
-          )
+          Operation.create(calc_date: calc_date)
         end
-
-        #
-        # 運用情報を生成する
-        #
-        # @param [Result::Data::SingleDay] calc_date 和暦日（計算値）
-        #
-        # @return [Result::Operation::Bundle] 運用情報
-        #
-        def self.create_operation(calc_date: Result::Data::SingleDay.new)
-          first_day = calc_date.month.first_day.western_date
-          operation_history = Operation.specify_history(western_date: first_day)
-
-          operation_month = create_operation_month(operation_history: operation_history)
-
-          Result::Operation::Bundle.new(
-            operated: !operation_history.invalid?, month: operation_month, original: calc_date
-          )
-        end
-        private_class_method :create_operation
-
-        #
-        # 月履歴集約情報を生成する
-        #
-        # @param [Operation::MonthHistory] operation_history 変更履歴（月）
-        #
-        # @return [Result::Operation::Month::Bundle] 月履歴集約情報
-        #
-        def self.create_operation_month(operation_history: Operation::MonthHistory.new)
-          return Result::Operation::Month::Bundle.new if operation_history.invalid?
-
-          parent_operation_history = Operation.specify_history_by_id(
-            id: operation_history.parent_id
-          )
-
-          Result::Operation::Month::Bundle.new(
-            current: create_operation_month_history(operation_history: operation_history),
-            parent: create_operation_month_history(operation_history: parent_operation_history)
-          )
-        end
-        private_class_method :create_operation_month
-
-        #
-        # 月別履歴情報を生成する
-        #
-        # @param [Operation::MonthHistory] operation_history 変更履歴（月）
-        #
-        # @return [Result::Operation::Month::History] 月別履歴情報
-        #
-        def self.create_operation_month_history(operation_history: Operation::MonthHistory.new)
-          return Result::Operation::Month::History.new if operation_history.invalid?
-
-          annotations = create_annnotations(operation_history: operation_history)
-
-          reference = operation_history.reference
-          Result::Operation::Month::History.new(
-            id: operation_history.id, western_date: operation_history.western_date.format,
-            page: reference.page, number: reference.number, annotations: annotations
-          )
-        end
-        private_class_method :create_operation_month_history
-
-        #
-        # 注釈情報を生成する
-        #
-        # @param [Operation::MonthHistory] operation_history 変更履歴（月）
-        #
-        # @return [Array<Result::Operation::Month::Annotation>] 注釈
-        #
-        def self.create_annnotations(operation_history: Operation::MonthHistory.new)
-          annotations = []
-          operation_history.annotations.each do |annotation|
-            annotations.push(
-              Result::Operation::Month::Annotation.new(
-                description: annotation.description,
-                note: annotation.note
-              )
-            )
-          end
-
-          annotations
-        end
-        private_class_method :create_annnotations
+        private_class_method :get_operation
       end
     end
   end
